test: 💍 add test for format document
